Move migrators into a subpackage.
