+++ internal/oobmigration/oobmigrators/migrators_test.go
@@ -1,4 +1,4 @@
-package oobmigration
+package oobmigrators
 
 import (
 	"context"
